Split off Transaction TransactionMap VersionedValue
