+++ h2/src/main/org/h2/mvstore/db/MVSpatialIndex.java
@@ -17,9 +17,9 @@ import org.h2.index.IndexType;
 import org.h2.index.SpatialIndex;
 import org.h2.index.SpatialTreeIndex;
 import org.h2.message.DbException;
-import org.h2.mvstore.tx.TransactionStore.Transaction;
-import org.h2.mvstore.tx.TransactionStore.TransactionMap;
-import org.h2.mvstore.tx.TransactionStore.VersionedValue;
+import org.h2.mvstore.tx.Transaction;
+import org.h2.mvstore.tx.TransactionMap;
+import org.h2.mvstore.tx.VersionedValue;
 import org.h2.mvstore.tx.TransactionStore.VersionedValueType;
 import org.h2.mvstore.rtree.MVRTreeMap;
 import org.h2.mvstore.rtree.MVRTreeMap.RTreeCursor;
